+++ test/teacher/components/molecules/ProfileSection.test.jsx
@@ -1,6 +1,6 @@
 import { describe, it, expect } from "vitest";
 import { render, screen } from "@testing-library/react";
-import { ProfileSection } from "../../../src/domains/teacher/components/molecules/ProfileSection";
+import { ProfileSection } from "../../../../src/domains/teacher/components/molecules/ProfileSection";
 
 describe("ProfileSection", () => {
   it("renders the title and children correctly", () => {
